move python scripts into py
