+++ stringdecomposer/py/standard_logger.py
